# event-platform-backend/swagger-corrected.yaml
openapi: 3.0.0
info:
  title: Event Platform API
  version: 1.0.0
  description: >
    API для платформы управления мероприятиями.
    Авторизация осуществляется через JWT, который сохраняется в httpOnly cookie после входа.
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  schemas:
...
    Event:
      type: object
      properties:
        event_id:
          type: integer
        photo_url:
          type: string
        name:
          type: string
        description:
          type: string
        limited:
          type: integer
        views:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        creator_tag:
          type: string
...
security:
  - cookieAuth: []
paths:
...
  /events:
    get:
      summary: Получить список мероприятий
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Номер страницы
        - in: query
          name: limited
          schema:
            type: integer
          description: Количество элементов на странице
        - in: query
          name: search
          schema:
            type: string
          description: Поисковый запрос
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          description: Дата начала фильтра
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          description: Дата окончания фильтра
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [start_date, end_date, views]
          description: Поле сортировки
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [ASC, DESC]
          description: Порядок сортировки
        - in: query
          name: minViews
          schema:
            type: integer
          description: Минимальное количество просмотров
        - in: query
          name: maxViews
          schema:
            type: integer
          description: Максимальное количество просмотров
      responses:
        '200':
          description: Список мероприятий с пагинацией
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      totalPages:
                        type: integer
                      limited:
                        type: integer

    post:
      summary: Создать мероприятие
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                limited:
                  type: integer
                photo:
                  type: string
                  format: binary
      responses:
        '201':
          description: Мероприятие создано
...
  /events/{id}:
    get:
      summary: Получить мероприятие по ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Мероприятие
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'

    put:
      summary: Обновить мероприятие
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                start_date:
                  type: string
                  format: date-time
                end_date:
                  type: string
                  format: date-time
                limited:
                  type: integer
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Мероприятие обновлено

    delete:
      summary: Удалить мероприятие
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Мероприятие удалено
...